requirements.txt
Django==4.2.5
mysqlclient==2.1.1
python-decouple==3.8
djangorestframework==3.14.0
Scrapy==2.11.0
requests==2.31.0